refactor(settings): GoogleTab extrahiert (H4 Schritt 5)
Reiner Move-Refactor: der ~440-Zeilen-Google-Block (Konto/Status,
Absender, Drive-Sync inkl. Konflikte/Import/Reconnect/Kompaktierung,
Google-Kalender) wandert aus open_settings_dialog in eine eigene
GoogleTab-Klasse. Verhaltensgleich, inkl. der H5-Hintergrund-Worker
(runner.run, winfo_exists-Guards, Persistenz im fn) unveraendert.
